fix(clippy): fix 12 clippy-1.95 errors in zenjpeg
Fixes for the clippy-1.95 lints that turned red after the toolchain
bump (independently of the magetypes fix in 9a0b8e66):

- manual_checked_ops → use `.checked_div` in:
    - encode/config.rs restart-interval overhead math
    - decode/pipeline.rs horizontal-padding downsampled_w
    - decode/mod.rs wave_buf memory sizing
- unnecessary u16→u16 casts in encode/symbol_stream.rs
- unused `mut` on `cfg` in codec.rs decode fallback
- wrong_self_convention: rename `to_pixels_into(&mut self, …)` to
  `write_pixels_into(&mut self, …)` (pub(in crate::decode); 2 call
  sites updated)
- collapsible_if in decode/parser/output.rs fast-path dispatch,
  decode/pipeline.rs horizontal padding, codec.rs cached_header init
- collapsible_match on `ThreadingPolicy::LimitOrSingle`/APP2 MPF arms
- useless vec![] in color/xyb.rs unit test
- dead-code on `decode_streaming_into` when built without `zencodec`:
  gate with `cfg(all(feature = "decoder", feature = "zencodec"))`

No functional change. Both CI clippy invocations
(`-p zenjpeg --lib --tests --benches` with and without the extended
feature set) now pass under `-D warnings`, and `cargo fmt -- --check`
is clean.

+++ zenjpeg/src/encode/symbol_stream.rs
@@ -149,7 +149,7 @@ pub(crate) fn block_to_symbols(
     let dc_cat = crate::entropy::category(dc_diff);
     if dc_cat > 0 {
         let extra = crate::entropy::additional_bits_with_cat(dc_diff, dc_cat);
-        stream.push(dc_class, dc_cat, extra as u16, dc_cat);
+        stream.push(dc_class, dc_cat, extra, dc_cat);
     } else {
         stream.push(dc_class, 0, 0, 0);
     }
@@ -166,7 +166,7 @@ pub(crate) fn block_to_symbols(
             }
             let ac_cat = crate::entropy::category(ac);
             let extra = crate::entropy::additional_bits_with_cat(ac, ac_cat);
-            stream.push(ac_class, (run << 4) | ac_cat, extra as u16, ac_cat);
+            stream.push(ac_class, (run << 4) | ac_cat, extra, ac_cat);
             run = 0;
         }
     }
